pyproject.toml
[project]
name = "i3e-partner"
version = "0.1.0"
description = "A tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-playwright>=0.7.0",
]
